# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Attendance Monitoring Dashboard API","version":"1.0.0"},"paths":{"/upload":{"post":{"tags":["upload"],"summary":"Upload Files","operationId":"upload_files_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_files_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UploadResponseItem"},"type":"array","title":"Response Upload Files Upload Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/files/":{"get":{"tags":["files"],"summary":"List Files","operationId":"list_files_files__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UploadedFileListItem"},"type":"array","title":"Response List Files Files  Get"}}}}}},"delete":{"tags":["files"],"summary":"Delete Files","operationId":"delete_files_files__delete","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/files/{file_id}":{"get":{"tags":["files"],"summary":"Get File Detail","operationId":"get_file_detail_files__file_id__get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"integer","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadedFileDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_files_upload_post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_upload_files_upload_post"},"DeleteRequest":{"properties":{"file_ids":{"items":{"type":"integer"},"type":"array","title":"File Ids"}},"type":"object","required":["file_ids"],"title":"DeleteRequest"},"DeleteResponse":{"properties":{"deleted_count":{"type":"integer","title":"Deleted Count"}},"type":"object","required":["deleted_count"],"title":"DeleteResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UploadResponseItem":{"properties":{"id":{"type":"integer","title":"Id"},"filename":{"type":"string","title":"Filename"},"uploaded_at":{"type":"string","format":"date-time","title":"Uploaded At"},"total_rows":{"type":"integer","title":"Total Rows"}},"type":"object","required":["id","filename","uploaded_at","total_rows"],"title":"UploadResponseItem"},"UploadedFileDetail":{"properties":{"id":{"type":"integer","title":"Id"},"filename":{"type":"string","title":"Filename"},"uploaded_at":{"type":"string","format":"date-time","title":"Uploaded At"},"header_order":{"items":{"type":"string"},"type":"array","title":"Header Order"},"rows":{"items":{"type":"object"},"type":"array","title":"Rows"}},"type":"object","required":["id","filename","uploaded_at","header_order","rows"],"title":"UploadedFileDetail"},"UploadedFileListItem":{"properties":{"id":{"type":"integer","title":"Id"},"filename":{"type":"string","title":"Filename"},"uploaded_at":{"type":"string","format":"date-time","title":"Uploaded At"},"total_rows":{"type":"integer","title":"Total Rows"}},"type":"object","required":["id","filename","uploaded_at","total_rows"],"title":"UploadedFileListItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}